CSS - iPhone fullscreen fix accidental drag of home button
The home button appears at the bottom of screen if the device has no real home button.
So if you try to drag the timeline you might drag the home button instead.
This fix moves the controlbar higher up based on Safari's env variables - only present if the home button is on screen.