Fullscreen: Detect existing fullscreen
If some other element (ideally body!) is already in fullscreen
we do not need to initiate fullscreen mode.
Instead we can just use CSS to make the player fullscreen.

This helps in Safari where the fullscreen mode cannot be a result of Ajax.
So you cannot load player HTML that will go to fullscreen automatically.