YouTube Logo: Cut off smaller part of video: Have to cut off bigger part
It seems the "More videos" box is now higher up.
It seems it only shows if player is wider than 515px.
Or does it detect screen at and above 598px?
So for .is-tiny we only need to cut off the YouTube video title area.
So the issue is less bad in portrait orientation of mobile,
but landscape is still a problem.