Fix FV Player screen search performance
This was really slow: JOIN `wp_fv_player_videos` AS v on FIND_IN_SET(v.id, p.videos)
There were 4 other JOINs with wp_fv_player_videometa too, but they did not seem to match.

So it's better to just get the player IDs for the matching video IDs
and then put in the videometa counts (subtitles, cues, chapters, transcript)
via a separate query which still uses the same JOIN,
but it's limitted to the amount of results you have.

Downside is that you can't sort by these columns.

To fix that we will have to query the videos in different way in FV_Player_Db::query_videos