Freedom Video Player: CSS Update
The controls should be easy to get rid of.
In fullscreen we made it so that they only show if you hover bottom of the player.
But when not in fullscreen it's easy to move your mouse out of the player.

So when not in fullscreen hovering the player anywhere will show controls.