Revert: Editor: Convert Gutenberg shortcode block to FV Player
This way you can use the shortcode block to post [fvplayer] shortcode with extra attributes,
which would be removed by the block code.

c6f294edc11b5a5a93bda06014f1ffc5a8362021 attempted to fix it for display,
but it would still need work for editing.
And it did not work on all websites.