iOS recovery: Fix HLS recovery, possibly for iOS 18
It seems the play with events is not enough.
What happened to me is that after blocking the CDN domain the iPhone would play the video up until the last loaded segment.
Then it would say the video duration is not what it was, but only the time until the last segment.
So I detect that and trigger and error if the video got shorter for no reason.

Maybe this should be the one and only technique for detecting HLS failures on iPhone.