Security: wp_kses for lightboxed players: Allow scripts
Otherwise the sizing script with el.classList.add( 'is-wide' ) would show as visible text when using FV Player Pro video ads.
The reason is that while real playlist are forced to use the "slider" playlist style which does not use that script, we have no easy way of detecting that for FV Player Pro video ads.